python/packages/kagent-adk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kagent-adk"
version = "0.3.0"
description = "kagent-adk is an sdk for integrating adk agents with kagent"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
  "agentsts-adk >= 0.0.6",
  "agentsts-core >= 0.0.6",
  "kagent-core",
  "aiofiles>=24.1.0",
  "anyio>=4.9.0",
  "typer>=0.15.0",
  "uvicorn>=0.34.0",
  "openai>=1.72.0",
  "mcp>=1.12.0",
  "protobuf>=6",
  "anthropic[vertex]>=0.49.0",
  "fastapi>=0.115.1",
  "litellm>=1.74.3",
  "google-adk>=1.16.0",
  "google-genai>=1.21.1",
  "google-auth>=2.40.2",
  "httpx>=0.25.0",
  "pydantic>=2.5.0",
  "typing-extensions>=4.8.0",
  "jsonref>=1.1.0",
  "a2a-sdk>=0.3.1",
]

[tool.uv.sources]
kagent-core = {workspace = true}

[project.scripts]
kagent-adk = "kagent.adk.cli:run_cli"

[project.optional-dependencies]
test = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "httpx>=0.25.0",  
]
memory = [
  "psutil>=6.1.0",  # For memory monitoring
]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]

[tool.ruff]
extend = "../../pyproject.toml"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"